(chore) Restructure files.
